requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.108.0
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
httptools==0.6.1
idna==3.6
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
protobuf==4.25.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pynng==0.7.4
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
wcferry==39.0.12.0
wcfhttp==39.0.10.0
websockets==12.0
python-dotenv